pyproject.toml
[project]
name = "browser-recorder"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sherwyn Sen", email = "sen@lpixel.net"},
]
dependencies = ["playwright>=1.50.0", "ffmpeg-python>=0.2.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
